% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtree.R
\name{.add_modelinfo}
\alias{.add_modelinfo}
\title{Add model information to a personalised-model-ctree}
\usage{
.add_modelinfo(x, nodeids, data, model, coeffun)
}
\arguments{
\item{x}{constparty object.}

\item{nodeids}{node ids, usually the terminal ids.}

\item{data}{data.}

\item{model}{model.}

\item{coeffun}{function that takes the model object and returns the coefficients.
Useful when coef() does not return all coefficients (e.g. survreg).}
}
\value{
tree with added info. Class still to be added.
}
\description{
For internal use.
}
