% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.prepare_args}
\alias{.prepare_args}
\title{Prepare input for ctree/cforest from input of pmtree/pmforest}
\usage{
.prepare_args(model, data, zformula, control, ...)
}
\arguments{
\item{model}{model.}

\item{data}{an optional data frame.}

\item{zformula}{ormula describing which variable should be used for partitioning.}

\item{control}{ontrol parameters, see \code{\link[partykit]{ctree_control}}.}

\item{...}{other arguments.}
}
\value{
args to be passed to ctree/cforest.
}
\description{
Prepare input for ctree/cforest from input of pmtree/pmforest
}
