% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pmtree.R
\name{lmplot}
\alias{lmplot}
\title{Density plot for a given lm model}
\usage{
lmplot(mod, data = NULL, densest = FALSE, theme = theme_classic())
}
\arguments{
\item{mod}{A model of class lm.}

\item{data}{optional data frame. If NULL the data stored in mod is used.}

\item{densest}{should additional to the model density kernel density estimates
(see \code{\link[ggplot2]{geom_density}}) be computed?}

\item{theme}{A ggplot2 theme.}
}
\description{
Can be used on its own but is also useable as plotfun in 
\code{\link{node_pmterminal}}.
}
\examples{
inst/examples/ex-pmtree-methods.R

}
