% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmodel-methods.R
\name{objfun.pmodel_identity}
\alias{objfun.pmodel_identity}
\alias{logLik.pmodel_identity}
\title{Objective function of personalised models}
\usage{
\method{objfun}{pmodel_identity}(x, ...)

\method{logLik}{pmodel_identity}(object, add_df = 0, ...)
}
\arguments{
\item{x, object}{object of class pmodel_identity (obtained by \code{pmodel(..., fun = identity)}).}

\item{...}{additional parameters passed on to \code{\link{objfun}}.}

\item{add_df}{it is not very clear what the degrees of freedom are in personalised models.
 With this argument you can add/substract degrees of freedom at your convenience. Default
 is \code{0} which means adding up the degrees of freedom of all individual models.

For examples see \code{\link{pmodel}}.}
}
\description{
Get the contributions of an objective function (e.g. likelihood contributions)
and the sum thereof (e.g. log-Likelihood).
}
