% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{one_factor}
\alias{one_factor}
\title{Check if model has only one factor covariate.}
\usage{
one_factor(object)
}
\arguments{
\item{object}{model.}
}
\value{
Returns TRUE if model has a single factor covariate, FALSE otherwise.
}
\description{
See https://stackoverflow.com/questions/50504386/check-that-model-has-only-one-factor-covariate/50514499#50514499
}
