% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmodel.R
\name{pmodel}
\alias{pmodel}
\title{Personalised model}
\usage{
pmodel(x = NULL, object = NULL, newdata = NULL, OOB = TRUE,
  fun = coef)
}
\arguments{
\item{x}{cforest object or matrix of weights.}

\item{object}{model object. If NULL the model in \code{x$info$object} is used.}

\item{newdata}{new data. If NULL cforest learning data is used. Ignored if \code{x} is a matrix.}

\item{OOB}{In case of using the learning data, should patient similarities be
computed out of bag?}

\item{fun}{function to apply on the personalised model before returning. The
default \code{coef} returns a matrix of personalised coefficients. For returning
the model objects use \code{identity}.}
}
\value{
depends on fun.
}
\description{
Compute personalised models from cforest object.
}

