% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtree.R
\name{pmtree}
\alias{pmtree}
\title{Compute model-based forest from model.}
\usage{
pmtree(object, data = NULL, zformula = ~., control = ctree_control(),
  coeffun = coef, ...)
}
\arguments{
\item{object}{a model object.}

\item{data}{data. If NULL (default) the data from the model object are used.}

\item{zformula}{formula describing which variable should be used for partitioning.
Default is to use all variables in data that are not in the model (i.e. \code{~ .}).}

\item{control}{control parameters, see \code{\link[partykit]{ctree_control}}.}

\item{coeffun}{function that takes the model object and returns the coefficients. 
Useful when \code{coef()} does not return all coefficients (e.g. \code{survreg}).}

\item{...}{additional parameters passed on to model fit such as weights.}
}
\value{
ctree object
}
\description{
Input a parametric model and get a forest.
}
\examples{
if(require("TH.data") & require("survival")) {
  ## base model
  bmod <- survreg(Surv(time, cens) ~ horTh, data = GBSG2, model = TRUE)
  survplot(bmod)
  
  ## partitioned model
  tr <- pmtree(bmod)
  plot(tr, terminal_panel = node_pmterminal(tr, plotfun = survplot, 
                                            confint = TRUE))
  summary(tr)
  summary(tr, node = 1:2)
  
  logLik(bmod)
  logLik(tr)
}
}
