% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtree.R
\name{print.pmtree}
\alias{print.pmtree}
\title{Print method for pmtree}
\usage{
\method{print}{pmtree}(x, node = NULL, FUN = NULL,
  digits = getOption("digits") - 4L, footer = TRUE, ...)
}
\arguments{
\item{x}{pmtree object.}

\item{node}{node number, if any.}

\item{FUN}{formatinfo function.}

\item{digits}{number of digits.}

\item{footer}{should footer be included?}

\item{...}{further arguments passed on to prin.party.}
}
\value{
print
}
\description{
Print pmtree
}

