% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objfun.R
\name{rss}
\alias{rss}
\alias{rss.default}
\title{Residual sum of squares}
\usage{
rss(object, ...)

\method{rss}{default}(object, ...)
}
\arguments{
\item{object}{model object.}

\item{...}{passed on to specific methods.}
}
\value{
sum of the squared residuals.
}
\description{
Returns the sum of the squared residuals for a given object.
}
\examples{
## example from ?lm
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- lm(weight ~ group)
rss(lm.D9)
}
