% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pmtree.R
\name{survplot}
\alias{survplot}
\title{Survival plot for a given survreg model}
\usage{
survplot(mod, data = NULL, theme = theme_classic())
}
\arguments{
\item{mod}{A model of class survreg.}

\item{data}{optional data frame. If NULL the data stored in mod is used.}

\item{theme}{A ggplot2 theme.}
}
\value{
RETURN DESCRIPTION
}
\description{
Can be used on its own but is also useable as plotfun in 
\code{\link{node_pmterminal}}.
}
\examples{
if(require("survival")) {
  survplot(survreg(Surv(futime, fustat) ~ ecog.ps + rx, ovarian))
}

}
