% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp.R
\name{varimp.pmforest}
\alias{varimp.pmforest}
\title{Variable Importance for pmforest}
\usage{
\method{varimp}{pmforest}(object, nperm = 1L, OOB = TRUE,
  risk = function(x, ...) -objfun(x, sum = TRUE, ...), ...)
}
\arguments{
\item{object}{DESCRIPTION.}

\item{nperm}{the number of permutations performed.}

\item{OOB}{a logical determining whether the importance is computed from the 
out-of-bag sample or the learning sample (not suggested).}

\item{risk}{the risk to be evaluated. By default the objective function 
(e.g. log-Likelihood) is used.}

\item{...}{passed on to \code{\link[partykit]{varimp.cforest}}.}
}
\value{
A vector of ‘mean decrease in accuracy’ importance scores.
}
\description{
See \code{\link[partykit]{varimp.cforest}}.
}
