% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtree-methods.R
\name{logLik.pmtree}
\alias{logLik.pmtree}
\title{Extract log-Likelihood}
\usage{
\method{logLik}{pmtree}(object, dfsplit = 0, newdata = NULL, weights = NULL, perm = NULL, ...)
}
\arguments{
\item{object}{pmtree object.}

\item{dfsplit}{degrees of freedom per selected split.}

\item{newdata}{an optional new data frame for which to compute the sum of 
objective functions.}

\item{weights}{weights.}

\item{perm}{the number of permutations performed (see \code{\link[partykit]{varimp}}).}

\item{...}{ignored.}
}
\value{
Returns an object of class \code{\link[stats]{logLik}}.
}
\description{
Extract sum of log-Likelihood contributions of all terminal nodes. By default the degrees
of freedom from the models are used but optionally degrees of freedom for splits
can be incorporated.
}
\seealso{
\code{\link{objfun.pmtree}} for the sum of contributions to the
objective function (not the same when partitioning linear models \code{\link[stats]{lm}})
}
