% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmforest.R
\name{pmforest}
\alias{pmforest}
\title{Compute model-based forest from model.}
\usage{
pmforest(object, data = NULL, zformula = ~., ntree = 2,
  control = ctree_control(lookahead = TRUE, mincriterion = 0, saveinfo =
  FALSE, ...), ...)
}
\arguments{
\item{object}{a model object.}

\item{data}{data. If NULL (default) the data from the model object are used.}

\item{zformula}{formula describing which variable should be used for partitioning.
Default is to use all variables in data that are not in the model (i.e. \code{~ .}).}

\item{ntree}{number of trees.}

\item{control}{control parameters, see \code{\link[partykit]{ctree_control}}.}

\item{...}{additional parameters passed on to model fit such as weights.}
}
\value{
cforest object
}
\description{
Input a parametric model and get a forest.
}

