% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtree.R
\name{pmtree}
\alias{pmtree}
\title{Compute model-based forest from model.}
\usage{
pmtree(object, data = NULL, zformula = ~., control = ctree_control(), ...)
}
\arguments{
\item{object}{a model object.}

\item{data}{data. If NULL (default) the data from the model object are used.}

\item{zformula}{formula describing which variable should be used for partitioning.
Default is to use all variables in data that are not in the model (i.e. \code{~ .}).}

\item{control}{control parameters, see \code{\link[partykit]{ctree_control}}.}

\item{...}{additional parameters passed on to model fit such as weights.}
}
\value{
ctree object
}
\description{
Input a parametric model and get a forest.
}

