% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtree-methods.R
\name{print.pmtree}
\alias{print.pmtree}
\alias{summary.pmtree}
\alias{print.summary.pmtree}
\alias{coef.pmtree}
\title{Methods for pmtree}
\usage{
\method{print}{pmtree}(
  x,
  node = NULL,
  FUN = NULL,
  digits = getOption("digits") - 4L,
  footer = TRUE,
  ...
)

\method{summary}{pmtree}(object, node = NULL, ...)

\method{print}{summary.pmtree}(x, digits = 4, ...)

\method{coef}{pmtree}(object, node = NULL, ...)
}
\arguments{
\item{x}{object.}

\item{node}{node number, if any.}

\item{FUN}{formatinfo function.}

\item{digits}{number of digits.}

\item{footer}{should footer be included?}

\item{...}{further arguments passed on to \code{\link[partykit]{print.party}}.}

\item{object}{object.}
}
\value{
print
}
\description{
Print and summary methods for pmtree objects.
}
